requirements.txt
pytumblr2
python-dotenv
httpx[http2]
pyyaml
